refactored tests directory
