+++ tests/test_agents/test_bandit/__init__.py
@@ -1,0 +1,6 @@
+from tests.test_agents.test_bandit.test_cb_agents import TestCBAgent  # noqa
+from tests.test_agents.test_bandit.test_data_bandits import TestDataBandit  # noqa
+from tests.test_agents.test_bandit.test_mab_agents import TestMABAgent  # noqa
+from tests.test_agents.test_bandit.test_multi_armed_bandits import (
+    TestMultiArmedBandit,  # noqa
+)
