+++ tests/test_agents/test_classical/__init__.py
@@ -1,0 +1,1 @@
+from tests.test_agents.test_classical.test_agents import test_qlearning, test_sarsa
