+++ tests/test_agents/test_deep/__init__.py
@@ -1,0 +1,8 @@
+from tests.test_agents.test_deep.test_a2c import test_a2c, test_a2c_cnn
+from tests.test_agents.test_deep.test_ddpg import test_ddpg
+from tests.test_agents.test_deep.test_dqn import TestDQN
+from tests.test_agents.test_deep.test_dqn_cnn import TestDQNCNN
+from tests.test_agents.test_deep.test_ppo1 import test_ppo1, test_ppo1_cnn
+from tests.test_agents.test_deep.test_sac import test_sac
+from tests.test_agents.test_deep.test_td3 import test_td3
+from tests.test_agents.test_deep.test_vpg import test_vpg, test_vpg_cnn
