--- tests/test_bandit/__init__.py
@@ -1,5 +1,0 @@
-from tests.test_bandit.test_bandit_trainer import test_bandit_trainer  # noqa
-from tests.test_bandit.test_cb_agents import TestCBAgent  # noqa
-from tests.test_bandit.test_data_bandits import TestDataBandit  # noqa
-from tests.test_bandit.test_mab_agents import TestMABAgent  # noqa
-from tests.test_bandit.test_multi_armed_bandits import TestMultiArmedBandit  # noqa
