--- tests/test_classical/__init__.py
@@ -1,2 +1,0 @@
-from tests.test_classical.test_agents import TestAgents
-from tests.test_classical.test_common import TestModels, TestTrainer
