--- tests/test_classical/test_common/__init__.py
@@ -1,2 +1,0 @@
-from tests.test_classical.test_common.test_models import TestModels
-from tests.test_classical.test_common.test_trainer import TestTrainer
