--- tests/test_deep/test_common/__init__.py
@@ -1,8 +1,0 @@
-from tests.test_deep.test_common.test_logger import test_loggers
-from tests.test_deep.test_common.test_trainer import (
-    test_load_params,
-    test_off_policy_trainer,
-    test_on_policy_trainer,
-    test_save_params,
-)
-from tests.test_deep.test_common.test_utils import TestUtils
