+++ tests/test_trainers/__init__.py
@@ -1,0 +1,8 @@
+from tests.test_trainers.test_bandit_trainer import test_bandit_trainer
+from tests.test_trainers.test_classical_trainer import test_classical_trainer
+from tests.test_trainers.test_deep_trainer import (
+    test_load_params,
+    test_off_policy_trainer,
+    test_on_policy_trainer,
+    test_save_params,
+)
