+++ tests/test_trainers/test_deep_trainer.py
